pyproject.toml
[project]
name = "langc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "hub>=3.0.1",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
]
